changed valuescard params
