+++ src/lib/data.ts
@@ -48,6 +48,30 @@ const sucursalesData: Sucursal[] = [
   // Agrega más sucursales con sus colores...
 ];
 
+const valuesicea = [
+  {
+    initial: 'I',
+    title: 'Innovación',
+    description: 'Buscamos constantemente nuevas formas de mejorar nuestros productos y procesos',
+  },
+  {
+    initial: 'C',
+    title: 'Calidad',
+    description: 'Mantenemos los más altos estándares en todos nuestros productos y servicios',
+  },
+  {
+    initial: 'E',
+    title: 'Excelencia',
+    description: 'Nos esforzamos por superar las expectativas en cada proyecto que emprendemos',
+  },
+  {
+    initial: 'A',
+    title: 'Alimentación',
+    description: 'Nos dedicamos a nutrir y alimentar de manera responsable y sostenible',
+  },
+];
+export const getValuesIcea = () => valuesicea;
+
 // Función SÍNCRONA para obtener una sucursal.
 // No necesita ser 'async' porque no espera a nada.
 export const getSucursalBySlug = (slug: string): Sucursal | undefined => {
